requirements.txt
aniso8601==6.0.0
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-Migrate==3.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2-binary==2.9.1
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==2.0.1
python-dotenv==0.19.2
pycryptodome==3.3.1
MarkupSafe==1.1.1
lazy-object-proxy==1.4.0
python-jose-cryptodome==1.3.2
pylint==2.3.1